\name{ICweights}
\alias{ICweights}
\title{IC weights functions for AIC and BIC weights}
\usage{
ICweights
loglikelihoodsc(-3,-5)
n_pars e.g.ec(4,5)
IC_type='AIC','BIC' or 'AICc'
sample size e.g. 1000
}
\description{
ICweights functions for AIC and BIC weights.
}
\examples{
ICweights(c(-3,-5),c(4,5),IC_type='AIC',1000)
}
